requirements.txt
Flask==3.0.3
python-docx==1.1.0
Jinja2==3.1.4
Werkzeug==3.0.3
itsdangerous==2.2.0
click==8.1.7
MarkupSafe==2.1.5